+++ common/src/main/java/whocraft/tardis_refined/constants/ModMessages.java
@@ -81,7 +81,7 @@ public class ModMessages {
     public static String ITEM_GROUP = "itemGroup." + TardisRefined.MODID;
 
     /*Messages*/
-
+    public static String SPAWN_TARDIS_DIMENSION_FAIL = message("spawn_tardis_dimension_fail");
 
     public static String MSG_EXTERIOR_COOLDOWN = message("exterior_cooldown");
     public static String MSG_KEY_BOUND = message("key_bound");