+++ changelog.md
@@ -2,6 +2,9 @@
 
 ![TARDIS Refined](https://wiki.tardisrefined.net/TARDIS-Refined-Wiki/tardis_refined_v2_1.png)
 
+#### Items
+- Added Grown Tardis Item.
+
 #### Changes
 - Amethyst Screwdriver tooltip now is colored gray.